Generalize the PrairieTest end-exam JWT helper
